btrfs-progs: sync ctree.c from kernel
This copies the current copy of ctree.c into btrfs-progs.  There are a
few modifications, they are all hi-lighted by a "MODIFIED" tag in the
code.

- The btrfs_path_cachep is a static struct to avoid needing to call
  btrfs_ctree_init to be able to use stuff.
- There is a single page optimization in the search that is not relevant
  in btrfs-progs.
- We have a few warnings around trans->transaction, but we don't have
  this distinction in btrfs-progs yet.
- Obvious include cleanups

The rest is a straight copy, the only "modifications" that I didn't note
were in ctree.h to clean up some of the inline helpers that use
different functions.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>