btrfs-progs: add a free_root_extent_buffers helper
Our CI started failing a bunch because I accidentally introduced an
extent buffer leak.  This is because we haphazardly have ->commit_roots
used in btrfs-progs, and they get freed when the transaction commits and
then they're cleared out.  In the kernel we make sure to free all this
when we free the root, but we don't have the same thing in btrfs-progs.
Fix this by bringing over the free_root_extent_buffers helper and use
this for free'ing up all the roots.  This brings us inline with the
kernel more and eliminates the extent buffer leak.

Reviewed-by: Anand Jain <anand.jain@oracle.com>
Signed-off-by: Josef Bacik <josef@toxicpanda.com>