btrfs-progs: update add_root_to_dirty_list to match the kernel
In the kernel our add_root_to_dirty_list function uses the
BTRFS_ROOT_DIRTY bit to see if we need to add the root to the dirty
list, and then places the extent roots at the back of the list to make
sure we're not updating the extent root over and over again as we update
the delayed refs.  Update this function to match the kernel version.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>