btrfs-progs: sync btrfs_bin_search from the kernel
In the kernel we have just one function instead of the
generic_bin_search which is called by btrfs_bin_search.  Update this to
match the kernel.

I had to modify this slightly as we have an optimization around the item
being in a single page in the kernel that doesn't belong in btrfs-progs.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>