btrfs-progs: sync search_leaf from the kernel
This is a helper for btrfs_search_slot, along with it's helper
search_for_key_slot it's used as an optimization for the binary search.
Sync these two helpers to enable use to sync btrfs_search_slot from the
kernel.  This will throw an unused error until we sync
btrfs_search_slot.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>