btrfs-progs: don't memset in btrfs_release_path
It is common to call btrfs_release_path and re-search for something with
the same path.  We use the different flags in the path to control how we
search, and we expect those flags to be preserved until we unset them.
Remove the memset from btrfs_release_path, which fixes various issues
I was seeing in the cases where we use ->search_for_extension.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>