btrfs-progs: remove btrfs_prev_leaf usage int btrfs check
When we're searching for a key and it is greater than the key that
exists, we will never end up at slot[0] in the next leaf, we will always
end up at slot nritems in the leaf that has our previous item.  The
btrfs_prev_leaf() usage here is trying to handle the case that our
previous item is in the leaf behind us, but the way search works we will
never end up there.  Remove these usages of btrfs_prev_leaf, and handle
the case of not finding the key we're looking for properly.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>